[Update] routes.rb内にadmin/genre/destroyアクションのルーティング追加

+++ app/controllers/admin/genres_controller.rb
@@ -30,11 +30,15 @@ class Admin::GenresController < Admin::ApplicationController
       redirect_to request.referer
     end
   end
-  
+
   def destroy
     genre = Genre.find(params[:id])
-    genre.destroy
-    redirect_to admin_genres_path, notice: "ジャンルの削除に成功しました。"
+    if genre.posts.empty?
+      genre.destroy
+      redirect_to admin_genres_path, notice: "ジャンルの削除に成功しました。"
+    else
+      redirect_to admin_genres_path, alert: "ジャンルに紐づいた記事が存在するため削除できません。"
+    end
   end
 
   private
